requirements.txt
django
djangorestframework==3.6.4
requests
pandas
gunicorn
